requirements.txt
psycopg2-binary==2.9.9
faker==19.12.0
python-dateutil==2.8.2
pandas==2.0.3
